remove env check
