+++ src/lib/constants.js
@@ -131,50 +131,11 @@ export const BETTER_COMMENTS_TAGS = [
 ];
 
 /**
- * Required npm packages for Salesforce development
- */
-export const REQUIRED_PACKAGES = [
-  '@salesforce/cli',
-  'prettier',
-  '@prettier/plugin-xml',
-  'prettier-plugin-apex',
-];
-
-/**
- * Required SF CLI plugins
- */
-export const REQUIRED_SF_PLUGINS = [
-  '@salesforce/sfdx-scanner',
-  'code-analyzer',
-];
-
-/**
  * Global state keys used by the extension
  */
 export const STATE_KEYS = {
-  PACKAGES_CHECKED: `${EXTENSION_ID}.packages-checked`,
-  SF_PLUGINS_CHECKED: `${EXTENSION_ID}.sf-plugins-checked`,
   WORKSPACE_INITIALIZED: `${EXTENSION_ID}.workspace-initialized`,
   THEME_SET: `${EXTENSION_ID}.has-theme-set`,
   BETTER_COMMENTS_SET: `${EXTENSION_ID}.has-better-comments-set`,
   PROMPTED_AUTO_UPDATE: `${EXTENSION_ID}.promptedForAutoUpdate`,
-  ENV_CHECK_COMPLETED: `${EXTENSION_ID}.env-check-completed`,
-};
-
-/**
- * Minimum required versions
- */
-export const MIN_VERSIONS = {
-  NODE: 18,
-  JAVA: 11,
-};
-
-/**
- * External URLs for documentation and downloads
- */
-export const EXTERNAL_URLS = {
-  JAVA_SETUP: 'https://developer.salesforce.com/docs/platform/sfvscode-extensions/guide/java-setup.html',
-  JAVA_DOWNLOAD: 'https://www.oracle.com/java/technologies/downloads/',
-  NODE_DOWNLOAD: 'https://nodejs.org/',
-  SALESFORCE_CLI: 'https://developer.salesforce.com/tools/salesforcecli',
 };
