+++ src/services/index.js
@@ -2,9 +2,5 @@
  * Services module exports
  */
 
-export * as environmentService from './environment.js';
-export * as statusBarService from './status-bar.js';
-export * as packageService from './packages.js';
-export * as pluginService from './sf-plugins.js';
 export * as settingsService from './settings.js';
 export * as sfdxService from './sfdx.js';
